Automation - CEPH-83574741-RBD-mirror:Removal of failed back image should reflect in secondary
