+++ suites/quincy/rbd/tier-2_rbd_mirror_regression.yaml
@@ -156,12 +156,13 @@ tests:
 
   - test:
       name: test image delete from secondary after multiple promote and demote
-      module: test_83574741.py
+      module: test-image-delete-from-secondary.py
       clusters:
         ceph-rbd1:
           config:
             imagesize: 2G
             io-total: 200M
+            repeat_count: 1
       polarion-id: CEPH-83574741
       desc: Verify that deleting primary image also delete the secondary image
